Implement @idiebrain's changes
These include bringing the mock resource up to speed with the actual implentation that supports the predeployed annotation on resources, and adding tests for cases where something can be predeployed and is marked as predeployed, and the inverse. Also a test for if a resource is marked as predeployed but can't be predeployed.